Breakup App.vue and router-view
Instead of having the page header as part of the App.vue file
I've moved it to the NestsMap.vue view. This means each page will
consist of only the stuff to the right of the sidebar, i.e. the
router-view is everything to the right of the sidebar. This means
I can use the page to coordinate component communication.

Shrink size of sidebar a little.

Issue #1: for now I'm traversing the component tree to send events
all the to App.vue. I'll need to change this once I include Vuex.